[CCUBE-256][KT] Pull out FileListItem into a component of it's own for future extraction
